+++ src/document/quote_paragraph.cpp
@@ -1,0 +1,5 @@
+#include "quote_paragraph.hpp"
+
+Paragraph::Level QuoteParagraph::level(void) const {
+	return Paragraph::Level::Quote;
+}