+++ src/document/quote_paragraph.hpp
@@ -1,0 +1,13 @@
+#ifndef MDP_QUOTE_PARAGRAPH_HPP
+#define MDP_QUOTE_PARAGRAPH_HPP
+
+#include "paragraph.hpp"
+
+class QuoteParagraph : public Paragraph {
+	public:
+		QuoteParagraph() : Paragraph() {}
+		QuoteParagraph(Paragraph * p) : Paragraph(p) {}
+		virtual Level level(void) const;
+};
+
+#endif // MDP_QUOTE_PARAGRAPH_HPP